requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
propcache==0.2.0
Pygments==2.18.0
requests==2.32.3
rich==13.9.4
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.17.1
